requirements/base.txt
# Base requirements for running this project in any environment.
Django==1.5.1
django_compressor==1.3
django-filter==0.6
django_localflavor_us==1.0
docutils==0.10
#Pillow==2.0.0
#psycopg2==2.5
python-geohash==0.8.4
requests==0.14.2